moved dimension stuff back into TableClass, use flatMap in get_dimensions
